requirements.txt
APScheduler==3.6.3
async-timeout==4.0.3
azure-core==1.30.1
azure-identity==1.15.0
azure-keyvault-secrets==4.8.0
cachetools==4.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
configparser==6.0.1
idna==3.6
pymongo==4.6.2
python-telegram-bot==13.7
pytz==2024.1
redis==5.0.2
requests==2.31.0
setuptools==69.1.1
six==1.16.0
telegram==0.0.1
tornado==6.4
tzlocal==5.2
urllib3==1.26.18
